+++ example/lib/global_error_data.dart
@@ -1,6 +1,11 @@
 import 'package:coder_matthews_extensions/coder_matthews_extensions.dart';
 
 class GlobalErrorData extends ErrorData {
-  GlobalErrorData(
-      {required super.message, required super.title, required super.exception, super.controllerSource, super.onError});
+  GlobalErrorData({
+    required super.message,
+    required super.title,
+    required super.exception,
+    super.controllerSource,
+    super.onError,
+  });
 }