+++ lib/src/utilities/type_definitions.dart
@@ -1,1 +1,6 @@
+import 'package:http/http.dart';
+
+import '../exceptions/exceptions.dart';
+
 typedef OnProgress = void Function(int current, int total);
+typedef OnHttpResponse = ControllerException? Function(Response response, Type source);